+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- feat: source TypeScript test command from gg\_lang catalog; use gg\_lang project detection
+
 ## [2.0.1] - 2026-05-04
 
 ## [2.0.0] - 2026-04-20
@@ -188,6 +194,7 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_test/compare/2.0.1...HEAD
 [2.0.1]: https://github.com/inlavigo/gg_test/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/inlavigo/gg_test/compare/1.1.8...2.0.0
 [1.1.8]: https://github.com/inlavigo/gg_test/compare/1.1.7...1.1.8